Fix routing back from onboarding (#78)
* Onboarding back button works, contains alternative code commented out as well

* Navigating flow from page to page works, but there are two warning messages for missing dependencies on the useEffect

* Removed warnings by forcing them to turn off

* Made modifica
tions such that the onboarding page now contains a cancel button, there's a toas
t warning against using the back button, and that it is possible to track where
the user previously came from. However, the previous page when navigating from r
ide summary is registered as auth. Need to fix this.

* Fixed the bug where the site is recording the previous page incorrectly for ridesummary page

* Removed console log messages

* Fixed issue with empty page when triggering the back button from ride summary in carpool

* Standardized the URL stored into the localStorage so that it doesnt contain the forward slash in front

Co-authored-by: Henry Qin <hq5@rice.edu>